// src/compositions/cricket/utils/broadcastPro/playerRanking/types.ts
export interface BroadcastProTripleStat {
  label1: string;
  value1: string;
  label2: string;
  value2: string;
  label3: string;
  value3: string;
}

export type BroadcastProPlayerRankingLayout =
  | "featured"
  | "gridTop5"
  | "gridPerformances";

export type BroadcastProPlayerRankBadgePlacement = "left" | "right";